requirements.txt
backports.lzma==0.0.8
certifi==2017.7.27.1
chardet==3.0.4
idna==2.5
protobuf==3.3.0
PyYAML==3.12
requests==2.18.3
six==1.10.0
urllib3==1.22
